refactor extend typescript support to dialog action, event and form
